+++ listener/upload_test.go
@@ -225,8 +225,8 @@ func TestUpdateSystemRepos1(t *testing.T) {
 
 	systemID := 5
 	rhAccountID := 1
-	database.Db.Create(models.SystemRepo{RhAccountID: rhAccountID, SystemID: systemID, RepoID: 1})
-	database.Db.Create(models.SystemRepo{RhAccountID: rhAccountID, SystemID: systemID, RepoID: 2})
+	database.Db.Create(models.SystemRepo{RhAccountID: int64(rhAccountID), SystemID: int64(systemID), RepoID: 1})
+	database.Db.Create(models.SystemRepo{RhAccountID: int64(rhAccountID), SystemID: int64(systemID), RepoID: 2})
 
 	repos := []string{"repo1", "repo10", "repo20"}
 	repoIDs, nReposAdded, err := ensureReposInDB(database.Db, repos)
@@ -248,10 +248,10 @@ func TestUpdateSystemRepos2(t *testing.T) {
 
 	systemID := 5
 	rhAccountID := 1
-	database.Db.Create(models.SystemRepo{RhAccountID: rhAccountID, SystemID: systemID, RepoID: 1})
-	database.Db.Create(models.SystemRepo{RhAccountID: rhAccountID, SystemID: systemID, RepoID: 2})
+	database.Db.Create(models.SystemRepo{RhAccountID: int64(rhAccountID), SystemID: int64(systemID), RepoID: 1})
+	database.Db.Create(models.SystemRepo{RhAccountID: int64(rhAccountID), SystemID: int64(systemID), RepoID: 2})
 
-	nAdded, nDeleted, err := updateSystemRepos(database.Db, rhAccountID, systemID, []int{})
+	nAdded, nDeleted, err := updateSystemRepos(database.Db, rhAccountID, systemID, []int64{})
 	assert.Nil(t, err)
 	assert.Equal(t, int64(0), nAdded)
 	assert.Equal(t, int64(2), nDeleted)
